refactor: cleanup dataset command module files
